add save report to file.
