+++ UrlStatus/Program.cs
@@ -1,8 +1,5 @@
 ﻿using System;
 using System.Windows.Forms;
-using log4net.Appender;
-using log4net.Config;
-using log4net.Layout;
 
 namespace UrlStatus
 {
@@ -16,8 +13,6 @@ namespace UrlStatus
         {
             Application.ThreadException += Application_ThreadException;
             AppDomain.CurrentDomain.UnhandledException += CurrentDomain_UnhandledException;
-
-            // LoadFileAppender();
 
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
@@ -36,40 +31,10 @@ namespace UrlStatus
 
         static void LogUnhandledException(object exceptionobj)
         {
-            LogManager.WriteLog("程序异常终止:" + exceptionobj);
+            Exception ex = (Exception)exceptionobj;
+            MessageBox.Show(ex.Message);
+            LogManager.Error(string.Format("LogUnhandledException:{0}.", ex.Message), ex);
         }
 
-        /// <summary>
-        /// 初始化日志文件配置
-        /// </summary>
-        static void LoadFileAppender()
-        {
-            string currentPath = AppDomain.CurrentDomain.BaseDirectory;
-
-            log4net.Repository.Hierarchy.Hierarchy hier =
-                log4net.LogManager.GetRepository() as log4net.Repository.Hierarchy.Hierarchy;
-
-            RollingFileAppender fileAppender = new RollingFileAppender();
-            fileAppender.Name = "RollingFileAppender";
-            fileAppender.File = "logs/log-";
-            fileAppender.AppendToFile = true;
-            fileAppender.RollingStyle = RollingFileAppender.RollingMode.Composite;
-            fileAppender.LockingModel = new log4net.Appender.RollingFileAppender.MinimalLock();
-            fileAppender.MaxSizeRollBackups = -1;
-            fileAppender.MaximumFileSize = "2MB";
-            fileAppender.DatePattern = "'APP-'yyyyMMdd'.log'";
-            fileAppender.StaticLogFileName = false;
-            fileAppender.Encoding = System.Text.Encoding.UTF8;
-
-            PatternLayout patternLayout = new PatternLayout();
-            patternLayout.ConversionPattern = "%date [%thread] %-5level%n %message%n %exception%n";
-            patternLayout.ActivateOptions();
-
-            fileAppender.Layout = patternLayout;
-
-            fileAppender.ActivateOptions();
-
-            BasicConfigurator.Configure(fileAppender);
-        }
     }
 }
